+++ tests/tests.spec.ts
@@ -1,5 +1,6 @@
 import { COLORS, HEX3, CMYK, FUNCTIONS, ColorProps, CMYKProps, ColorTranslator, Harmony } from './data/data';
-import { HSLObject } from '../src/@types';
+import { HSLObject, ColorInput } from '../src/@types';
+import { ERRORS } from '../src/constants';
 
 //Iterate over the colors
 COLORS.forEach((item: ColorProps): void => {
@@ -38,45 +39,83 @@ COLORS.forEach((item: ColorProps): void => {
     });
 });
 
-// Test HSL Objects with percentages
+// Test HSL Objects
 const hslTo = ['rgb', 'rgbObject', 'rgba', 'rgbaObject', 'hex', 'hexObject', 'hexa', 'hexaObject'];
 
 COLORS.forEach((item: ColorProps): void => {
 
-    describe(`HSL Object with percentages`, (): void => {
+    describe('HSL Object tests', (): void => {
 
         const hsl = { ...item.hslObject as HSLObject };
         const hsla = { ...item.hslaObject as HSLObject };
-
-        hsl.s += '%';
-        hsl.l += '%';
-        hsla.s += '%';
-        hsla.l += '%';
+        const hslPercentages = { ...hsl, s: hsl.s + '%', l: hsl.l + '%' };
+        const hslaPercentages = { ...hsla, s: hsla.s + '%', l: hsla.l + '%' };
+        const hslBigHue = { ...hsl, h: hsl.h + 360 };
+        const hslaBigHue = { ...hsl, h: hsl.h + 360 };
+        const hslLowHue = { ...hsl, h: hsl.h - 360 };
+        const hslaLowHue = { ...hsl, h: hsl.h - 360 };
 
         hslTo.forEach((prop: keyof ColorProps): void => {
 
-            it(`${JSON.stringify(hsl)} => ${JSON.stringify(item[prop])}`, (): void => {
+            const functionCall = FUNCTIONS[prop].func;
+            const cssProps = FUNCTIONS[prop].css;
 
-                const functionCall = FUNCTIONS[prop].func;
-                const cssProps = FUNCTIONS[prop].css;
+            it(`HSL Objects with percentages ${JSON.stringify(hslPercentages)} => ${JSON.stringify(item[prop])}`, (): void => {
 
                 if (cssProps) {
-                    expect(functionCall(hsl, true)).toBe(item[prop]);
+                    expect(functionCall(hslPercentages, true)).toBe(item[prop]);
                 } else {
-                    expect(functionCall(hsl, false)).toMatchObject(item[prop]);
+                    expect(functionCall(hslPercentages, false)).toMatchObject(item[prop]);
                 }
 
             });
 
-            it(`${JSON.stringify(hsla)} => ${JSON.stringify(item[prop])}`, (): void => {
-
-                const functionCall = FUNCTIONS[prop].func;
-                const cssProps = FUNCTIONS[prop].css;
+            it(`HSLA Objects with percentages ${JSON.stringify(hslaPercentages)} => ${JSON.stringify(item[prop])}`, (): void => {
 
                 if (cssProps) {
-                    expect(functionCall(hsla, true)).toBe(item[prop]);
+                    expect(functionCall(hslaPercentages, true)).toBe(item[prop]);
                 } else {
-                    expect(functionCall(hsla, false)).toMatchObject(item[prop]);
+                    expect(functionCall(hslaPercentages, false)).toMatchObject(item[prop]);
+                }
+
+            });
+
+            it(`HSL Objects big hue ${JSON.stringify(hslBigHue)} => ${JSON.stringify(item[prop])}`, (): void => {
+
+                if (cssProps) {
+                    expect(functionCall(hslBigHue, true)).toBe(item[prop]);
+                } else {
+                    expect(functionCall(hslBigHue, false)).toMatchObject(item[prop]);
+                }
+
+            });
+
+            it(`HSLA Objects big hue ${JSON.stringify(hslaBigHue)} => ${JSON.stringify(item[prop])}`, (): void => {
+
+                if (cssProps) {
+                    expect(functionCall(hslaBigHue, true)).toBe(item[prop]);
+                } else {
+                    expect(functionCall(hslaBigHue, false)).toMatchObject(item[prop]);
+                }
+
+            });
+
+            it(`HSL Objects low hue ${JSON.stringify(hslLowHue)} => ${JSON.stringify(item[prop])}`, (): void => {
+
+                if (cssProps) {
+                    expect(functionCall(hslLowHue, true)).toBe(item[prop]);
+                } else {
+                    expect(functionCall(hslLowHue, false)).toMatchObject(item[prop]);
+                }
+
+            });
+
+            it(`HSLA Objects low hue ${JSON.stringify(hslaLowHue)} => ${JSON.stringify(item[prop])}`, (): void => {
+
+                if (cssProps) {
+                    expect(functionCall(hslaLowHue, true)).toBe(item[prop]);
+                } else {
+                    expect(functionCall(hslaLowHue, false)).toMatchObject(item[prop]);
                 }
 
             });
@@ -145,6 +184,12 @@ CMYK.forEach((item: CMYKProps): void => {
         it(`${item.cmykint} => ${hex}`, (): void => {
             expect(ColorTranslator.toHEX(item.cmykint)).toBe(hex);
         });
+        it(`${JSON.stringify(item.cmykObject)} => ${hex}`, (): void => {
+            expect(ColorTranslator.toHEX(item.cmykObject)).toBe(hex);
+        });
+        it(`${JSON.stringify(item.cmykIntObject)} => ${hex}`, (): void => {
+            expect(ColorTranslator.toHEX(item.cmykIntObject)).toBe(hex);
+        });
     });
 });
 
@@ -155,13 +200,27 @@ describe('ColorTranslator blending tests', (): void => {
         const from = '#FF0000';
         const to = '#0000FF';
 
-        const blend1 = ColorTranslator.getBlendHEX(from, to, 3);
-        const blend2 = ColorTranslator.getBlendHEX(from, to, 6);
-        const blend3 = ColorTranslator.getBlendHEX(from, to, 9);
+        const blendFunctions = [
+            { colorFn: ColorTranslator.toHEX, blendFn: ColorTranslator.getBlendHEX },
+            { colorFn: ColorTranslator.toHEXA, blendFn: ColorTranslator.getBlendHEXA },
+            { colorFn: ColorTranslator.toRGB, blendFn: ColorTranslator.getBlendRGB },
+            { colorFn: ColorTranslator.toRGBA, blendFn: ColorTranslator.getBlendRGBA },
+            { colorFn: ColorTranslator.toHSL, blendFn: ColorTranslator.getBlendHSL },
+            { colorFn: ColorTranslator.toHSLA, blendFn: ColorTranslator.getBlendHSLA }
+        ];
 
-        expect(blend1).toMatchObject([ '#FF0000', '#7F007F', '#0000FF' ]);
-        expect(blend2).toMatchObject([ '#FF0000', '#CC0033', '#990066', '#660099', '#3300CC', '#0000FF' ]);
-        expect(blend3).toMatchObject([ '#FF0000', '#DF001F', '#BF003F', '#9F005F', '#7F007F', '#5F009F', '#3F00BF', '#1F00DF', '#0000FF' ]);
+        const result1 = [ '#FF0000', '#800080', '#0000FF' ];
+        const result2 = [ '#FF0000', '#CC0033', '#990066', '#660099', '#3300CC', '#0000FF' ];
+
+        blendFunctions.forEach((obj): void => {
+
+            const r1 = result1.map(c => obj.colorFn(c));
+            const r2 = result2.map(c => obj.colorFn(c));
+
+            expect(obj.blendFn(from, to, r1.length)).toMatchObject(r1);
+            expect(obj.blendFn(from, to, r2.length)).toMatchObject(r2);
+
+        });
         
         // Check default steps
         expect(ColorTranslator.getBlendHEX(from, to)).toMatchObject(ColorTranslator.getBlendHEX(from, to, 5));
@@ -171,29 +230,64 @@ describe('ColorTranslator blending tests', (): void => {
 
 });
 
-describe('ColorTranslator harmony tests', (): void => {
+describe('ColorTranslator harmony tests', (): void => {    
 
-    it('Harmony deep equals', (): void => {
+    const colorFunctions = [
+        ColorTranslator.toHEX,
+        ColorTranslator.toHEXA,
+        ColorTranslator.toRGB,
+        ColorTranslator.toRGBA,
+        ColorTranslator.toHSL,
+        ColorTranslator.toHSLA
+    ];
 
-        const base = '#FF0000';
-        
-        const results = [
-            ['#FF0000', '#FF0080', '#FF8000'],
-            ['#FF0000', '#00FFFF'],
-            ['#FF0000', '#0080FF', '#00FF80'],
-            ['#FF0000', '#00FF00', '#0000FF'],
-            ['#FF0000', '#FF00FF', '#00FF00', '#00FFFF'],
-            ['#FF0000', '#8000FF', '#80FF00', '#00FFFF']
-        ];
+    const base = '#FF0000';
+    
+    const results = [
+        ['#FF0000', '#FF0080', '#FF8000'],
+        ['#FF0000', '#00FFFF'],
+        ['#FF0000', '#0080FF', '#00FF80'],
+        ['#FF0000', '#00FF00', '#0000FF'],
+        ['#FF0000', '#FF00FF', '#00FF00', '#00FFFF'],
+        ['#FF0000', '#8000FF', '#80FF00', '#00FFFF']
+    ];
 
-        Object.keys(Harmony).forEach((harmony, index): void => {
+    Object.keys(Harmony).forEach((harmony, index): void => {
 
-            const colors = ColorTranslator.getHarmony(base, Harmony[harmony as Harmony]);
-            
-            expect(colors).toMatchObject(results[index]);
+        colorFunctions.forEach((fn): void => {
 
-        });
-        
+            const b = fn(base);
+            const r = results.map(colors => colors.map(color => fn(color)));
+
+            it(`Harmony deep equals: ${harmony} for ${b} => ${JSON.stringify(r[index])}`, (): void => {
+
+                const colors = ColorTranslator.getHarmony(b, Harmony[harmony as Harmony]);
+                expect(colors).toMatchObject(r[index]);
+
+            });            
+
+        });  
+
+    });
+
+});
+
+describe('Wrong inputs', (): void => {
+
+    it('Wrong string', (): void => {
+
+        expect(() => {
+            ColorTranslator.toHEX('AAA');
+        }).toThrowError(ERRORS.NOT_ACCEPTED_STRING_INPUT);
+
+    });
+
+    it('Wrong object', (): void => {
+
+        expect(() => {
+            ColorTranslator.toHEX({p: 100, h: 20, t: 360} as unknown as ColorInput);
+        }).toThrowError(ERRORS.NOT_ACCEPTED_OBJECT_INPUT);
+
     });
 
 });